feat: allow reuse of Dask schedulers
Previously, the Dask client was always shut down when leaving the global
settings context. Because cluster creation and client connection happened
only in the scheduler’s __init__, there was no way to restart them once
shut down.

This change introduces a method to restart a cluster and connect a new
client if the cluster was already shut down. As a result, the same
scheduler object can be reused for a second experiment in the same Python
process, allowing models to be reused across different global setting
contexts.